zocker-data/OCR/requirements.txt
Flask
Pillow
pytesseract
requests